refactor: FSD Phase 6a — 純粋な汎用フックを shared/lib/hooks へ移す (FRESTYLE-166)
ビジネス知識を持たない汎用フック 10 個(useAutoResizeTextarea /
useBackendHealth / useCopyToClipboard / useLocalStorage / useFormField /
useMobilePanelState / useFadeOnVisible / useSmoothReveal / useToast /
useToastContext)を shared/lib/hooks へ移す。features 層を切る前に、
どのプロジェクトでも使える純粋なフックを先に退避させる。

useNoteKeyboardShortcuts(n=ノート作成の意味づけ)・useSidebar(ログアウト
操作を含む)・useCommandPalette(アプリ固有のコマンド一覧に依存)は feature/
widgets 寄りなので対象外にした。

vi.mock の相対パス追従で 1 段の '../useToast' を取りこぼし、useAiSession /
useProfileEdit のテストが旧パスを指して 22 件落ちた。移動していないファイルの
テストでも、移動したフックを mock していれば追従が必要。絶対パスに直して解消。

既存テストの assert は変更していない。テスト件数 1184 で不変。
