+++ frontend/src/components/MessageInput.tsx
@@ -1,7 +1,7 @@
 import { useState, useRef, useEffect, KeyboardEvent, ChangeEvent } from 'react';
 import { ArrowUpIcon, PlusIcon } from '@heroicons/react/24/solid';
 import { XMarkIcon, PhotoIcon } from '@heroicons/react/24/outline';
-import { useAutoResizeTextarea } from '../hooks/useAutoResizeTextarea';
+import { useAutoResizeTextarea } from '@/shared/lib/hooks/useAutoResizeTextarea';
 import { AiChatRepository as aiChatRepository } from '@/entities/ai-chat';
 import type { AiAttachment, AiAttachmentFormat } from '@/entities/ai-chat';
