+++ frontend/src/hooks/useLoginPage.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import { useFormField } from './useFormField';
+import { useFormField } from '@/shared/lib/hooks/useFormField';
 import { AuthRepository as authRepository } from '@/entities/user';
 import { getApiError } from '@/shared/lib/classifyApiError';
