+++ frontend/src/pages/course-detail/ui/CourseDetailPage.tsx
@@ -23,9 +23,9 @@ import { useTeachingMaterialEditor } from '@/hooks/useTeachingMaterialEditor';
 import { useChapterResume } from '@/hooks/useChapterResume';
 import { useNextCourse } from '@/hooks/useNextCourse';
 import { useLessonProgress } from '@/hooks/useLessonProgress';
-import { useMobilePanelState } from '@/hooks/useMobilePanelState';
-import { useLocalStorage } from '@/hooks/useLocalStorage';
-import { useToast } from '@/hooks/useToast';
+import { useMobilePanelState } from '@/shared/lib/hooks/useMobilePanelState';
+import { useLocalStorage } from '@/shared/lib/hooks/useLocalStorage';
+import { useToast } from '@/shared/lib/hooks/useToast';
 import { CourseRepository } from '@/entities/course';
 import { DashboardRepository } from '@/entities/user';
 import { ImageUploadRepository } from '@/entities/user';
