+++ frontend/src/pages/courses/ui/CoursesListPage.tsx
@@ -16,7 +16,7 @@ import { CourseProgressBar } from '@/entities/course';
 import LanguageBadge from '@/shared/ui/LanguageBadge';
 import { FilterChip } from '@/shared/ui';
 import { useCourses } from '@/hooks/useCourses';
-import { useToast } from '@/hooks/useToast';
+import { useToast } from '@/shared/lib/hooks/useToast';
 import { COURSE_CATEGORIES, findCourseCategory } from '@/entities/course';
 import { COURSE_LANGUAGES } from '@/entities/course';
 import type { RootState } from '@/store';
