+++ frontend/src/shared/lib/hooks/__tests__/useAutoResizeTextarea.test.ts
