+++ frontend/src/shared/lib/hooks/__tests__/useCopyToClipboard.test.ts
