+++ frontend/src/shared/lib/hooks/__tests__/useFadeOnVisible.test.tsx
