+++ frontend/src/shared/lib/hooks/__tests__/useMobilePanelState.test.ts
