+++ frontend/src/shared/lib/hooks/__tests__/useSmoothReveal.test.ts
