+++ frontend/src/shared/lib/hooks/useAutoResizeTextarea.ts
