+++ frontend/src/shared/lib/hooks/useBackendHealth.ts
