+++ frontend/src/shared/lib/hooks/useCopyToClipboard.ts
