+++ frontend/src/shared/lib/hooks/useFadeOnVisible.ts
