+++ frontend/src/shared/lib/hooks/useMobilePanelState.ts
