+++ frontend/src/shared/lib/hooks/useSmoothReveal.ts
